0.0.4b: refactor and fix bugs
