+++ tailwind.config.ts
@@ -5,6 +5,14 @@ export default {
     "{routes,islands,components}/**/*.{ts,tsx}",
   ],
   darkMode: 'class',
+  safelist: [
+    'bg-red-500',
+    'bg-blue-500',
+    'bg-green-500',
+    'bg-purple-500',
+    'bg-yellow-500',
+    'bg-orange-500',
+  ],
   theme: {
     extend: {
       keyframes: {
